move examples to github_test package in examples_test.go
See #367 for discussion.
