+++ github/misc_test.go
@@ -46,20 +46,6 @@ func TestMarkdown(t *testing.T) {
 	}
 }
 
-func ExampleClient_Markdown() {
-	client := NewClient(nil)
-
-	input := "# heading #\n\nLink to issue #1"
-	opt := &MarkdownOptions{Mode: "gfm", Context: "google/go-github"}
-
-	output, _, err := client.Markdown(input, opt)
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	fmt.Println(output)
-}
-
 func TestListEmojis(t *testing.T) {
 	setup()
 	defer teardown()
